back/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
Faker==19.2.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
wrapt==1.15.0
